Add gates

+++ app/Http/Controllers/Admin/UsersController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\User;
 use App\Role;
+use Gate;
 use App\Http\Requests\UserAddRequest;
 use Illuminate\Http\Request;
 
@@ -34,6 +35,10 @@ class UsersController extends Controller
      */
     public function edit(User $user)
     {
+       if(Gate::denies('edit-users'))
+       {
+           return redirect()->route('admin.users.index');
+       }
         $roles = Role::All();
         return view('user.edit')->with([
             "user" => $user,
@@ -62,6 +67,10 @@ class UsersController extends Controller
      */
     public function destroy(User $user)
     {
+        if(Gate::denies('delete-users'))
+        {
+            return redirect()->route('admin.users.index');
+        }
         $user->roles()->detach();
         $user->delete();
         return redirect()->route('admin.users.index');
